fix history slider date formatting
The format string that we used to format the history dates on the
dashboard page was interpreted differently based on locale so the
dates were incorrect in the US (although probably OK in Europe).  In
the US we use "mm/dd/yyyy" where in Europe they use "dd/mm/yyyy" so in
the US the columns always showed "January".

It looks as if Ruby can handle ISO-8601-ish dates so "yyyy-mm-dd" is
less likely to be locale-dependent.