ensure that the dashboard history ajax update doesn't use the application layout
If the app in which rails_admin was installed had an application
layout (which most of them do) the dashboard history
update (i.e. dragging the history slider) would get rendered in that
layout which messes up the page.  Added a ":layout => false" to ensure
there's no layout and a spec to make sure it doesn't happen again.