take AR validations into account when deciding whether a field is required
Previously, a field's "required" or "optional" status in the
create/edit form was based only on that field's nullable status in the
database schema.  This led to cases where a field was indicated
"optional" but you'd get an error if you tried to save with that field
empty.  This commit adds checks for two kinds of validations:
ActiveModel::Validations::PresenceValidator and
ActiveModel::Validations::NumericalityValidator with
:allow_nil=>false.  There might well be other validations that cause
the fields to be required, but these are the two that I know about.