use padding instead of relative positioning for table headers
The problem with relative positioning is that it brings the elements
forward in the z-stack so the table headers were "higher" than the
menu of "other" tabs.  This made it hard to pick from that list since
the table headers would overlay it and interrup the hover, making the
menu disappear.  Padding gives roughly the same formatting, but the
table headers are below the menu in the z-stack.