rename histories table to rails_admin_histories
The problem with "histories" is that it's an obvious enough name that
it's likely to cause conflicts with existing tables.  We've already
seen one issue filed because of this.

This patch adds another migration to the existing generator to change
the name of the histories table, and changes the History class to use
that name.