add a helpful error message if the user hasn't migrated their db
It seems reasonable that people won't know to re-run the migration
generator so we want to give them a clue what went wrong.  The error
that comes back is database-specific but it always seems to have the
table name in it so we check to see if that's in the message and then
add our text to it, so even if we're wrong the user will see the whole
message.