remove 'require' from gemspec to work with warbler
JRuby/Warbler has a bug that caused apps to crash during
initialization if any of the gems' gemspec had a "require" statement.
See http://jira.codehaus.org/browse/JRUBY-5319 for more info.  For now
we're going to remove the "require" but we can add it back in when the
bug is fixed.  I think it's worth it to be able to run in a
jruby/warbler environment.