assume list view column width of 120px if not specified
When we group model attributes into "sets" for display on the list
view, we do the grouping by adding the width of the columns and
comparing that to the width of the page.  When the running total gets
wider than the page we put those columns into a set and start the
count again.  The columns always used to have a hard-coded default
width, but now we want to let the browser flow the list view so we
will soon remove those defaults.  120px is a pretty arbitrary number
but it seems to produce reasonable results, at least on the "baseball"
data set.