issue #362: specify link routes explicitly on model list page
We were specifying routes implicitly, i.e. passing hashes or
parameters to link_to and form_tag but not specifying the route.  This
works fine unless your routes file has the old "wildcard" route
declaration in it, in which case Rails builds a route based on the
current controller and action, which in this case ends up being
"/rails_admin/main/list" which isn't what we want.

The fix is to always specify the route explicitly so the wildcard
route never gets used.

Thanks to "qianthinking" on github for pointing us to a solution.