Added optional support for models that have a state machine.
The point is that if a model has a state field, we don't want to
set the state directly, rather we want to send an event to trigger
a transition which, in turn, has callbacks to do stuff.

All this would not happen if the state is set directly.

This change adds (not ideal) the event buttons next to the existing
history and delete buttons at the top of an edit page for a model.