validation refactoring.
It makes no sense to show non-nullable fields as required since saving the parent object would generate an exception. Of course you can catch it and treat it manually, but one shouldn't except RA to support anything but the uppest level of implementation, i.e. ActiveModel in that case. Validation for association is possible, so removed the exception for has_one. About belongs_to associations, it is possible to validate the child key or the child object. The 'correct' way to do it is child key, but as far as RA is concerned validating associated object should be just as fine. Also any validator that is not nullable should trigger required, not only NumericValidator.