remove ActiveRecord#rails_admin
, as it is fundamentally incompatible with models mass-configuration and https://github.com/sferik/rails_admin/blob/master/lib/rails_admin/config/base.rb#L78 configuration pattern.
Guarantying order and effective execution of both initializer and model's block in both prod and dev mode can't be met easily without destroying performances in dev mode (and complexifying loader by an order of magnitude, jeopardizing RA development).

So be it, it ain't worth it. rails_admin initializer is the king of the day.