Accept MySQL's 'utf8mb4' as a kind of UTF-8 for CSV export
MySQL's default UTF-8 encoding is limited to 3-byte characters and
the encoding it uses and reports for full 4-byte Unicode is called
'utf8mb4' but if the database reports this encoding, the CSV export
module would try to convert string encodings utf8mb4 -> utf8.
Obviously the former is not a valid encoding name in Ruby, but is
actually going to be UTF-8, so this change assumes it's already
UTF-8.