fix: save_and_edit should authorize the current object
was authorizing the class. Not sure why I'm the first person to see this.

Honestly, I was kind of surprised my `edit?` policy (using Pundit) was called twice.  It might be better to pass in a local to the partial since whether or not we can edit has already been determined