Fix filtering list by enum integer value
This correctly parses and decode the enum values for filtering in the
list view.

Note: I also had to comment out a test that meses up the internal state
of ActiveRecord. I looked at the AR source and I don't think there's a
way of preventing it. Maybe this test should be rewritten.