Rails 5: Corretly populate enum fields in form
This commit changes an order in which a type of enum field is determined.
Previously a generic enum field was preferred over ActiveRecord enum field,
however it caused AR enums to be recognized as generic enums.

Still, it worked correctly until Rails 5, where calling `safe_send` for enum
field returns a string value (while previously it returned numeric value).

Due to this change, `selected` parameter does not take the correct value and
on edit form, the enum field is always empty.

It does not cause a problem if enum field is correctly recognized as ActiveRecord enum.
Therefore, to fix the problem, I swapped the priority of enum fields.