Make colorpicker field use HTML5 native color picker
This removes the jquery.colorpicker.js dependency, reducing the overall
size of compiled assets.

The native color input element is builtin and so will have better a11y
support. It allows the browser, device, or a11y technology to present a
UI to the user to select a color. The value submitted by the form is a 6
digit hex, same as before.

Beyond the advantage of using a native UI element, this helps move
towards removing jQuery as a dependency or at least reducing its usage.

For additional details on the color input element, see:
https://developer.mozilla.org/en-US/docs/Web/HTML/Element/Input/color

Uses Rails color_field for rendering:
https://api.rubyonrails.org/classes/ActionView/Helpers/FormHelper.html#method-i-color_field

Refs #2893