Find-and-replace sferik owner with railsadminteam
Noticed that some of the links were broken because they pointed to the
repo with sferik as owner, but the repo has been transfered to
railsadminteam.

GitHub generally does a pretty good job of forwarding these along so the
links for issues/pull requests still worked, but references to
individual commits or files were broken.