Remove support for EOL Ruby 2.5
Ruby 2.5 went EOL 2021-03-31. It is no longer supported upstream and so
is not receiving bug fixes, including for security issues. Reduce the
testing resources on RailsAdmin.

https://www.ruby-lang.org/en/downloads/branches/

Dropping EOL Ruby versions allows the project to begin taking advantage
of newer syntax and library features.

TargetRubyVersion was removed from .rubocop.yml as this value is
determined automatically from the gemspec.