Move docs from wiki to the project repository itself
Moving the docs allows them to more easily be developed in a first-class
manner through pull requests. The docs can stay aligned with code
changes as they happen within the same commit.

For example, as the code gets refactored, docs may become incorrect or
out of date. With the docs in the repository, this can be caught during
development and review.

Contributors can provide reviewable documentaiton-only changes, even
small ones, thus improving the community experience.

An up to date version of the docs were cloned and copied using:

    git clone https://github.com/railsadminteam/rails_admin.wiki.git

The docs were left as-is as much as possible. If there are desired
changes, these can now be opened as future pull requests. Articles that
are linked to internally are not included.

Changes to the wiki articles:

- All previous wiki links have been converted to be compatible with
  GitHub flavored Markdown so links continue to work as before.

- Small formatting changes to be compliant with Prettier.

- References to *.coffee files were updated to *.js.

- Renamed index.md to home.md to follow conventions.

- Added an H1 header to all articles which was previously implicitly added by
  the wiki engine.

Once this change lands, the wiki can be removed so as to avoid confusion
for new users.