Remove JavaScript library jquery-pjax (#3435)
* Remove JavaScript library jquery-pjax

The library is no longer maintained and has not received a new release
since May 20, 2017. See https://github.com/defunkt/jquery-pjax

> Status of this project
>
> jquery-pjax is largely unmaintained at this point.

This also removes the Ruby gem rack-pjax which is also no longer
maintained and has not received a new release since Mar 24, 2019.

This is work towards the goal of dropping jQuery in favor of more
vanilla JavaScript solutions. As a dependency using jQuery, pjax was a
barrier toward this goal.

jquery-pjax was originally introduced in
645d678732eb550db2c5b250aa16f5ac5e9f8d62 (Nov 16, 2011). Presumably as a
page loading optimization, however it is not strictly necessary to
achieve the set of RailsAdmin features. After many years of improvements
in browser caching, loading, and JavaScript evaluation it is unclear if
this optimization is still necessary for a good UX.

Refs #2893