maintain 3.x deferred initializer block config
In 3.x the initializer config blocks are intentionally deferred to allow
referencing model classes without quoting. Here is a first pass at keeping that
behavior on top of ec03eb6f. It could probably use a little cleanup but here's
an idea of an approach for feedback.

See
https://github.com/railsadminteam/rails_admin/pull/3492#issuecomment-1064915362
for details.