Add whitespace after navigation icon
As discussed by mshibuya and codealchemy in [upgrade to Bootstrap v5](https://github.com/railsadminteam/rails_admin/pull/3455/files#r780533297) the rendering of icons should not be fixed with CSS but adding a whitespace between the icon and the text.

This change applies that solution to the icons in the sidebar navigation as well.