Remove support for end-of-life Ruby 2.6
Ruby 2.6 has been EOL since 2022-04-12. It is no longer receiving
security or bug fixes and its use is not supported. See:
https://www.ruby-lang.org/en/downloads/branches/

Removing Ruby 2.6 will reduce the testing and maintenance required for
this version. Workarounds can be removed. The RailsAdmin project can
begin adopting newer Ruby coding conventions when convenient.

This removes jruby-9.3 from CI testing as that represents Ruby 2.6.
jruby-9.4 was not added as that does not pass test suite.