When sorting models for navigation, fallback to model name
In my project, several STI models have the same label stored in their
translation file. This is because, from the user's perspective, they
represent the same abstract concept, even if the developers know
differently.

This can sometimes be an issue when rendering the main navigation. As
the labels are equal, the listed models can be displayed in either
order. It would be better to always display models in a single
consistent and deterministic order. This is especially an issue for my
project as the end-to-end tests rely on this order to assert correct
behavior.